requirements.txt
altair==5.2.0
beautifulsoup4==4.12.2
googlesearch_python==1.2.3
langchain==0.0.354
nsepython==2.8
numpy==1.23.4
pandas==1.5.2
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.28.1
yfinance==0.2.33
faiss_cpu
google-generativeai
